feat: unit tests
