+++ CadastroApi/CadastroApi/Application/Contato/Listar/ListarContatoQuery.cs
@@ -1,12 +1,11 @@
 ﻿using CadastroApi.Models;
 using MediatR;
 
-namespace CadastroApi.Application
+namespace CadastroApi.Application;
+
+public class ListarContatoQuery : IRequest<IEnumerable<Contato>>
 {
-    public class ListarContatoQuery : IRequest<IEnumerable<Contato>>
-    {
-        public string? DDD { get; set; }
-        public int? PageIndex { get; set; }
-        public int? PageSize { get; set; }
-    }
+    public string? DDD { get; set; }
+    public int? PageIndex { get; set; }
+    public int? PageSize { get; set; }
 }
