+++ CadastroApi/CadastroApi/Application/Contato/Listar/ListarContatoQueryHandler.cs
@@ -2,23 +2,22 @@
 using CadastroApi.Repository;
 using MediatR;
 
-namespace CadastroApi.Application
+namespace CadastroApi.Application;
+
+public class ListarContatoQueryHandler : IRequestHandler<ListarContatoQuery, IEnumerable<Contato>>
 {
-    public class ListarContatoQueryHandler : IRequestHandler<ListarContatoQuery, IEnumerable<Contato>>
+    private readonly IContatoRepository _contatoRepository;
+
+    public ListarContatoQueryHandler(IContatoRepository contatoRepository)
     {
-        private readonly IContatoRepository _contatoRepository;
+        _contatoRepository = contatoRepository;
+    }
 
-        public ListarContatoQueryHandler(IContatoRepository contatoRepository)
-        {
-            _contatoRepository = contatoRepository;
-        }
+    public async Task<IEnumerable<Contato>> Handle(ListarContatoQuery query, CancellationToken cancellationToken)
+    {
+        if(string.IsNullOrEmpty(query.DDD))
+            return await _contatoRepository.GetAllAsync(query.PageIndex, query.PageSize);
 
-        public async Task<IEnumerable<Contato>> Handle(ListarContatoQuery query, CancellationToken cancellationToken)
-        {
-            if(string.IsNullOrEmpty(query.DDD))
-                return await _contatoRepository.GetAllAsync(query.PageIndex, query.PageSize);
-
-            return await _contatoRepository.GetByDDDAsync(query.DDD, query.PageIndex, query.PageSize);
-        }
+        return await _contatoRepository.GetByDDDAsync(query.DDD, query.PageIndex, query.PageSize);
     }
 }
