+++ CadastroApi/CadastroApi/Application/Contato/Remover/RemoverContatoCommand.cs
@@ -1,14 +1,12 @@
 ﻿using MediatR;
 
-namespace CadastroApi.Application.RemoverContato
-{
-    public class RemoverContatoCommand : IRequest<Guid>
-    {
-        public Guid ContatoId { get; }
-        public RemoverContatoCommand(Guid id)
-        {
-            ContatoId = id;
-        }
-    }
+namespace CadastroApi.Application;
 
+public class RemoverContatoCommand : IRequest<Guid>
+{
+    public Guid ContatoId { get; }
+    public RemoverContatoCommand(Guid id)
+    {
+        ContatoId = id;
+    }
 }
