+++ CadastroApi/CadastroApi/Application/Usuario/Adicionar/AdicionarUsuarioCommand.cs
@@ -1,23 +1,21 @@
-﻿using CadastroApi.Application.Validators;
-using CadastroApi.Enums;
+﻿using CadastroApi.Enums;
 using FluentValidation;
 
-namespace CadastroApi.Application
-{
-    public class AdicionarUsuarioCommand
-    {
-        public string Nome { get; set; } = string.Empty;
-        public string Senha { get; set; } = string.Empty;
-        public TipoUsuarioPermissao? Permissao { get; set; }
+namespace CadastroApi.Application;
 
-        public void Validate()
+public class AdicionarUsuarioCommand
+{
+    public string Nome { get; set; } = string.Empty;
+    public string Senha { get; set; } = string.Empty;
+    public TipoUsuarioPermissao? Permissao { get; set; }
+
+    public void Validate()
+    {
+        var validator = new ModelUsuarioValidator();
+        var validationResult = validator.Validate(this);
+        if (!validationResult.IsValid)
         {
-            var validator = new ModelUsuarioValidator();
-            var validationResult = validator.Validate(this);
-            if (!validationResult.IsValid)
-            {
-                throw new ValidationException(validationResult.Errors);
-            }
+            throw new ValidationException(validationResult.Errors);
         }
     }
 }
