+++ CadastroApi/CadastroApi/Application/Usuario/Adicionar/AdicionarUsuarioCommandHandler.cs
@@ -1,30 +1,29 @@
 ﻿using CadastroApi.Models;
 using CadastroApi.Repository;
 
-namespace CadastroApi.Application
+namespace CadastroApi.Application;
+
+public class AdicionarUsuarioCommandHandler
 {
-    public class AdicionarUsuarioCommandHandler
+    private readonly IUsuarioRepository _usuarioRepository;
+
+    public AdicionarUsuarioCommandHandler(IUsuarioRepository usuarioRepository)
     {
-        private readonly IUsuarioRepository _usuarioRepository;
+        _usuarioRepository = usuarioRepository;
+    }
 
-        public AdicionarUsuarioCommandHandler(IUsuarioRepository usuarioRepository)
+    public async Task<Guid> Handle(AdicionarUsuarioCommand command, CancellationToken cancellationToken)
+    {
+        command.Validate();
+
+        var usuario = new Usuario
         {
-            _usuarioRepository = usuarioRepository;
-        }
+            Nome = command.Nome,
+            Permissao = command.Permissao.Value //Validate deve garantir que permissão não é null aqui
+        };
 
-        public async Task<Guid> Handle(AdicionarUsuarioCommand command, CancellationToken cancellationToken)
-        {
-            command.Validate();
+        await _usuarioRepository.AddUserAsync(usuario, command.Senha);
 
-            var usuario = new Usuario
-            {
-                Nome = command.Nome,
-                Permissao = command.Permissao.Value //Validate deve garantir que permissão não é null aqui
-            };
-
-            await _usuarioRepository.AddUserAsync(usuario, command.Senha);
-
-            return usuario.Id;
-        }
+        return usuario.Id;
     }
 }
