+++ CadastroApi/CadastroApi/Application/Usuario/Remover/RemoverUsuarioCommandHandler.cs
@@ -1,25 +1,24 @@
 ﻿using CadastroApi.Repository;
 
-namespace CadastroApi.Application
+namespace CadastroApi.Application;
+
+public class RemoverUsuarioCommandHandler
 {
-    public class RemoverUsuarioCommandHandler
+    private readonly IUsuarioRepository _usuarioRepository;
+
+    public RemoverUsuarioCommandHandler(IUsuarioRepository usuarioRepository)
     {
-        private readonly IUsuarioRepository _usuarioRepository;
+        _usuarioRepository = usuarioRepository;
+    }
 
-        public RemoverUsuarioCommandHandler(IUsuarioRepository usuarioRepository)
-        {
-            _usuarioRepository = usuarioRepository;
-        }
+    public async Task<Guid> Handle(RemoverUsuarioCommand command, CancellationToken cancellationToken)
+    {
+        var usuario = await _usuarioRepository.GetByIdAsync(command.UsuarioId);
+        if(usuario is null)
+            throw new KeyNotFoundException();
 
-        public async Task<Guid> Handle(RemoverUsuarioCommand command, CancellationToken cancellationToken)
-        {
-            var usuario = await _usuarioRepository.GetByIdAsync(command.UsuarioId);
-            if(usuario is null)
-                throw new KeyNotFoundException();
+        await _usuarioRepository.Delete(command.UsuarioId);
 
-            await _usuarioRepository.Delete(command.UsuarioId);
-
-            return usuario.Id;
-        }
+        return usuario.Id;
     }
 }
