+++ CadastroApi/CadastroApi/Repository/IUsuarioRepository.cs
@@ -1,10 +1,9 @@
 ﻿using CadastroApi.Models;
 
-namespace CadastroApi.Repository
+namespace CadastroApi.Repository;
+
+public interface IUsuarioRepository : IRepository<Usuario>
 {
-    public interface IUsuarioRepository : IRepository<Usuario>
-    {
-        Task AddUserAsync(Usuario usuario, string senha);
-        Task<Usuario?> GetUserAsync(string nome, string senha);
-    }
+    Task AddUserAsync(Usuario usuario, string senha);
+    Task<Usuario?> GetUserAsync(string nome, string senha);
 }
