+++ CadastroApi/UnitTest/Contato/RemoverContatoTests.cs
@@ -1,4 +1,4 @@
-﻿using CadastroApi.Application.RemoverContato;
+﻿using CadastroApi.Application;
 using CadastroApi.Controllers;
 using CadastroApi.Repository;
 using MediatR;
@@ -22,13 +22,13 @@ public class RemoverContatoTests
 
     [Fact]
     public async Task RemoverContato_InformadoContatoExistente_DeverRetornarOk()
-    {        
+    {
         var contatoId = Guid.NewGuid();
         var command = new RemoverContatoCommand(contatoId);
 
         _mediatorMock.Setup(m => m.Send(It.IsAny<RemoverContatoCommand>(), It.IsAny<CancellationToken>()))
                      .ReturnsAsync(contatoId);
-       
+
         var result = await _controller.RemoverContato(contatoId);
 
         var okResult = Assert.IsType<OkObjectResult>(result);
